"Refactored package structure and imports; removed interfaces IField and Game; renamed model classes to field.Field and game.Game"
