+++ src/main/scala/de/htwg/se/minesweeper/aview/MinesweeperGUI.scala
@@ -1,6 +1,6 @@
 package de.htwg.se.minesweeper.aview
 
-import de.htwg.se.minesweeper.controller.Controller
+import de.htwg.se.minesweeper.controller.controller.Controller
 import de.htwg.se.minesweeper.util.Observer
 import de.htwg.se.minesweeper.model.{Symbols, Status}
 import scala.swing._
