+++ src/main/scala/de/htwg/se/minesweeper/controller/ICommand.scala
@@ -1,4 +1,4 @@
-package de.htwg.se.minesweeper.interfaces
+package de.htwg.se.minesweeper.controller
 
 trait ICommand {
     def execute(): Unit
