+++ src/main/scala/de/htwg/se/minesweeper/factory/FieldFactory.scala
@@ -1,6 +1,7 @@
 package de.htwg.se.minesweeper.factory
 
-import de.htwg.se.minesweeper.model.{Field, Symbols}
+import de.htwg.se.minesweeper.model.Symbols
+import de.htwg.se.minesweeper.model.field.Field
 
 object FieldFactory {
   def createField(gridSize: Int): Field = {
